Replace .. links with baseurl calls
This still results in one hiccup on links that rely on redirects in es/lecciones.md: jekyll-redirect-from hardcodes the site.url in before the site.baseurl, so it will try to kick you back to http://programminghistorian.org. That's in the gem behavior, so it can't be easily worked around. Other than that, this fixes most of the url/baseurl problems.